New Snipping Tool with a new_UI and exectuable in the v1.02 Folder.
It will automatically save the image in the clipboard.

the polygon tool will do the same but with a white background,
transparent is not working on the clipboard, I will not fix that.
My Coworkers need it like that.

You can interrupt the Polygon by pressing on the last point a double click,
Esc or Enter Button.